requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.17.1
cssselect2==0.7.0
Django==4.2
fonttools==4.55.0
mysqlclient==2.2.6
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
PyMySQL==1.1.1
pyphen==0.15.0
python-dotenv==1.0.1
sqlparse==0.5.2
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2024.2
weasyprint==63.0
webencodings==0.5.1
zopfli==0.2.3.post1
